requirements.txt
pandas==2.0.01
numpy==1.24.3
seaborn==0.12.2
matplotlib==3.7.1
scikit-learn==1.2.2
scipy==1.10.1
imbalanced-learn==0.10.1
flake8==6.0.0
isort==5.12.0
haversine==2.8.0
python-dotenv==1.0.0
mlflow==2.3.0
numba==0.57.0
